requirements_pip.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
cycler==0.10.0
decorator==4.3.2
entrypoints==0.3
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pygraphviz==1.5
PyGSP
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==5.1
pyzmq==17.1.2
qtconsole==4.4.3
scikit-learn==0.20.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
